Move BottomSheetHandler to core-design-system
